Moved rectified linear unit activation to it's own file.
